+++ src/components/loginForm/LoginForm.tsx
@@ -1,8 +1,8 @@
+import { LoginData } from '@data/formData';
+import { userSchema } from '@data/formSchemas';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { setTokenToLocalStorage } from '@utils/auth';
-import { LoginData } from '@utils/formData';
-import { submitFormData } from '@utils/formHandler';
-import { userSchema } from '@utils/formSchemas';
+import { setTokenToLocalStorage } from '@middleware/authHandler';
+import { submitFormData } from '@middleware/formHandler';
 import React from 'react';
 import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
