+++ src/components/mock/mockMeetingsAll.tsx
@@ -1,6 +1,6 @@
-import { MeetingData, Meetings } from '@utils/dataTypes';
-import MockMeetingComponent from '@utils/mockMeetingComponent';
+import { MeetingData, Meetings } from '@data/dataTypes';
 import React from 'react';
+import MockMeetingComponent from './mockMeetingComponent';
 
 const MockMeetings = ({ meetings }: { meetings: Meetings }): JSX.Element => {
 	return (
