+++ src/components/utils/NewMeetingForm.tsx
@@ -1,13 +1,13 @@
+import { NewMeetingData } from '@data/formData';
+import { newMeetingSchema } from '@data/formSchemas';
 import { zodResolver } from '@hookform/resolvers/zod';
 import '@layouts/FormLayout.css';
-import { NewMeetingData } from '@utils/formData';
-import { submitFormData } from '@utils/formHandler';
+import { submitFormData } from '@middleware/formHandler';
 import {
 	calculateMinTime,
 	convertToUTCDateTime,
 	validateDayTime,
-} from '@utils/formHelperFunctions';
-import { newMeetingSchema } from '@utils/formSchemas';
+} from '@middleware/helpers/dateTimeHelper';
 import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router';
@@ -101,6 +101,11 @@ function NewMeetingForm() {
 				{errors.date && (
 					<span className='form-element__validation-prompt'>
 						{errors.date.message}
+					</span>
+				)}
+				{errorMessage && (
+					<span className='form-element__validation-prompt'>
+						{errorMessage}
 					</span>
 				)}
 			</div>
